smart-kyc/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pyyaml==6.0.1
python-multipart==0.0.6
pathway==0.8.0
pandas==2.1.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
requests==2.31.0
pillow==10.1.0
aiofiles==23.2.1
pydantic==2.5.0
python-dotenv==1.0.0
